requirements.txt
amqp==2.6.0
aniso8601==8.0.0
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.3
billiard==3.6.3.0
cffi==1.14.4
cfgv==3.1.0
click==7.1.2
cryptography==3.2.1
distlib==0.3.1
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
flask-restx==0.2.0
future==0.18.2
identify==1.4.25
importlib-metadata==1.7.0
iniconfig==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==4.6.11
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
nodeenv==1.4.0
packaging==20.9
pluggy==0.13.1
pre-commit==2.6.0
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==2.0.1
pylint==2.5.3
pylint-django==2.2.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.2.2
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
uWSGI==2.0.19.1
vine==1.3.0
virtualenv==20.0.28
Werkzeug==0.16.0
wrapt==1.12.1
zipp==3.1.0
